app/requirements.txt
alembic
#aniso8601
#blessings
#certifi
#chardet
#click
#curtsies
Flask
Flask-HTTPAuth
Flask-Migrate
Flask-RESTful
Flask-Script
Flask-SQLAlchemy
#greenlet
gunicorn
#idna
itsdangerous
Jinja2
#Mako
#pylibmc
python-dateutil
psycopg2
#pytz
PyYAML
requests
##six
SQLAlchemy
transliterate
#typing
urllib3
#uWSGI
#Werkzeug
